separate temp data
